// frontendmusic/src/types.ts
export type RootStackParamList = {
  Profile: undefined;
  EditProfile: undefined;
  MusicLanguage: undefined;
  StreamingQuality: undefined;
  DownloadQuality: undefined;
  Login: undefined;
  LikedSongsScreen: undefined;
  PlaylistsScreen: undefined;
  ArtistsFollowingScreen: undefined;
  YourLibraryScreen: undefined;
  DownloadsScreen: undefined;
  SongScreen: { song: { id: string; title: string; artist: string; image: string } };
  QueueScreen: undefined;
};